Add file size logging for tiles and resources
