+++ packages/client/src/utils.js
@@ -4,6 +4,17 @@ import url from 'url';
 import path from 'path';
 import crypto from 'crypto';
 
+// Formats a raw byte integer as a string
+export function formatBytes(int) {
+  let units = ['kB', 'MB', 'GB'];
+  let base = 1024;
+  let u = -1;
+
+  if (Math.abs(int) < base) return `${int}B`;
+  while (Math.abs(int) >= base && u++ < 2) int /= base;
+  return `${int.toFixed(1)}${units[u]}`;
+}
+
 // Returns a sha256 hash of a string.
 export function sha256hash(content) {
   return crypto
