+++ packages/core/src/install.js
@@ -4,18 +4,7 @@ import path from 'path';
 import https from 'https';
 import logger from '@percy/logger';
 import cp from 'child_process';
-import { ProxyHttpsAgent } from '@percy/client/utils';
-
-// Formats a raw byte integer as a string
-function formatBytes(int) {
-  let units = ['kB', 'MB', 'GB'];
-  let base = 1024;
-  let u = -1;
-
-  if (Math.abs(int) < base) return `${int}B`;
-  while (Math.abs(int) >= base && u++ < 2) int /= base;
-  return `${int.toFixed(1)}${units[u]}`;
-}
+import { ProxyHttpsAgent, formatBytes } from '@percy/client/utils';
 
 // Formats milleseconds as "MM:SS"
 function formatTime(ms) {
